Fixes varios

+++ src/main/java/Estado.java
@@ -1,5 +1,5 @@
 public class Estado {
 
     public double beneficio = 0;
-    public int suscriptores = 70000;
+    public long suscriptores = 70000;
 }
