+++ src/main/java/Globales.java
@@ -1,13 +1,13 @@
 public class Globales {
 
     public double modificadorVerano = 0;
-    public int modificadorMundial = 0;
+    public long modificadorMundial = 0;
     public double modificadorStrike = 0;
 
     /**
      * Para la funcion Funciones.medidoresAnuales
      */
-    public int aniosTranscurridos = 0;
+    public long aniosTranscurridos = 0;
     public double indiceSuscriptores = 0;
     public double indiceBeneficio = 0;
     public double indiceStrike = 0;
@@ -15,30 +15,30 @@ public class Globales {
     /**
      * Para los strikes
      */
-    public int finDeStrike = 0;
-    public int cantidadDeStrikes = 0;
+    public long finDeStrike = 0;
+    public long cantidadDeStrikes = 0;
 
     /**
      * Para los setUp de dias de semana y fin de semana
      */
-    public int periodistasDisponibles = 0;
-    public int partidosACubrir = 0;
-    public int partidosImportantes = 0;
-    public int partidosNormales = 0;
+    public long periodistasDisponibles = 0;
+    public long partidosACubrir = 0;
+    public long partidosImportantes = 0;
+    public long partidosNormales = 0;
 
     /**
      * Para las sumatorias o acumuladores
      */
-    public int sumatoriaPeriodistasDisponibles = 0;
-    public int sumatoriaPartidosImportantesACubrir = 0;
-    public int sumatoriaPartidosNormalesACubrir = 0;
+    public long sumatoriaPeriodistasDisponibles = 0;
+    public long sumatoriaPartidosImportantesACubrir = 0;
+    public long sumatoriaPartidosNormalesACubrir = 0;
 
-    public int sumatoriaPartidosImportantesTransmitidos = 0;
-    public int sumatoriaPartidosNormalesTransmitidos = 0;
+    public long sumatoriaPartidosImportantesTransmitidos = 0;
+    public long sumatoriaPartidosNormalesTransmitidos = 0;
 
-    public int sumatoriaVistasDiarias = 0;
-    public int sumatoriaPartidosACubrir = 0;
-    public int sumatoriaPeriodistasOciosos = 0;
+    public long sumatoriaVistasDiarias = 0;
+    public long sumatoriaPartidosACubrir = 0;
+    public long sumatoriaPeriodistasOciosos = 0;
 
     /**
      * Para cubrir los partidos
@@ -46,7 +46,7 @@ public class Globales {
     public double gananciaDiaria = 0;
     public double costoDiarioOperadores = 0;
     public double costoDiarioPeriodistas = 0;
-    public int vistasDiarias = 0;
-    public int partidosNoTransmitidosDiarios = 0;
-    public int gananciasPorVista = 10;
+    public long vistasDiarias = 0;
+    public long partidosNoTransmitidosDiarios = 0;
+    public long gananciasPorVista = 10;
 }
